services/api/requirements.txt
# Production dependencies - pinned for stability
# To upgrade: pip install pip-tools && pip-compile --upgrade requirements.in

fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
alembic==1.13.1
python-multipart==0.0.9
httpx==0.27.0
python-jose[cryptography]==3.3.0

# Testing dependencies
pytest==8.0.0
pytest-asyncio==0.23.5
asgi-lifespan==2.1.0

# Additional dependencies
aiosqlite==0.19.0
paho-mqtt==1.6.1

# Rate limiting
slowapi==0.1.9

# Auth (JWT + password hashing)
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Metrics
prometheus-client==0.20.0
